+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2024-12-21
+
 ### Added
 - nip25: Reactions
 - nip30: Custom Emoji
@@ -79,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/theborakompanioni/nostr-spring-boot-starter/compare/0.0.4...HEAD
+[Unreleased]: https://github.com/theborakompanioni/nostr-spring-boot-starter/compare/0.1.0...HEAD
+[0.1.0]: https://github.com/theborakompanioni/nostr-spring-boot-starter/releases/tag/0.1.0...0.0.4
 [0.0.4]: https://github.com/theborakompanioni/nostr-spring-boot-starter/releases/tag/0.0.4...0.0.3
 [0.0.3]: https://github.com/theborakompanioni/nostr-spring-boot-starter/releases/tag/0.0.3...0.0.2
 [0.0.2]: https://github.com/theborakompanioni/nostr-spring-boot-starter/releases/tag/0.0.2...0.0.1